# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AilaoDa ERP/CRM API
  version: 0.1.0
  description: >
    Seed OpenAPI contract for the current ERP/CRM API surface. This file is
    intentionally partial: it documents the active route families and a small
    set of high-value endpoints first, then must be expanded module by module
    before it is treated as the authoritative generated API contract.
  x-contract-status: seed
  x-contract-owner: engineering-readiness
  x-source-evidence:
    - backend/src/server.ts
    - backend/src/routes/auth.routes.ts
    - backend/src/routes/dashboard.routes.ts
    - backend/src/routes/customer.routes.ts
    - backend/src/routes/order.routes.ts
servers:
  - url: http://127.0.0.1:5001
    description: Local stable runtime
...
security:
  - bearerAuth: []
paths:
...
  /api/customers:
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Customer list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Customers]
      summary: Create a customer
      responses:
        '200':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List sales orders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Orders]
      summary: Create a sales order
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      additionalProperties: true
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
...